% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_statespace_var.R
\name{as_statespace_var}
\alias{as_statespace_var}
\title{Convert a VAR with p lags to a statespace representation with just one lag}
\usage{
as_statespace_var(betas, sigma)
}
\arguments{
\item{betas}{matrix of coefficients from a VAR.
Should have constant coefficient in the first column if included.
Dimensions should be k x (1 + k * p), or k x (k x p) if no constant.}

\item{sigma}{matrix of the var-cov of the residuals.
Should be k x k in dimension.}
}
\value{
matrix coefficients for VAR 1 representation
}
\description{
Convert a VAR with p lags to a statespace representation with just one lag
}
