% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Identify the main shock by targetting the forecast error
variance contribution in the frequency domain.}
\usage{
bootstrap(
  var,
  id_method = NULL,
  nboot = 500,
  n_ahead = 20,
  design = "recursive",
  method = "resample",
  wild_distr = "gaussian",
  bias_adjust = FALSE,
  lower_pctl = 0.16,
  upper_pctl = 0.84,
  ...
)
}
\arguments{
\item{var}{VAR to bootstrap}

\item{id_method}{function that identifies the Structural VAR}

\item{nboot}{number of bootstrap iteratons}

\item{n_ahead}{length of IRF horizon}

\item{design}{"recursive" or "fixed".
Controls how samples of the data are constructed.}

\item{method}{"resample" or "wild"}

\item{wild_distr}{distribution used for "wild" boostrap method.
Either "gaussian", "rademacher", or "mammen".}

\item{bias_adjust}{Boolean.  Set to TRUE to calculate the bias
and run the bootstrap again.}

\item{lower_pctl}{lower percentile of bootstraps returned in summary.
Defaults to 0.16 (for 68\% CI).}

\item{upper_pctl}{upper percentile of bootstraps returned in summary.
Defaults to 0.84 (for 68\% CI).}

\item{...}{Additional arguments passed to the function passed to id_method}
}
\value{
list of bootstrapped VARs and IRFs
}
\description{
Identify the main shock by targetting the forecast error
variance contribution in the frequency domain.
}
