% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe.R
\name{fe}
\alias{fe}
\title{Calculate the historical forecast error for a VAR out to a specified horizon}
\usage{
fe(var, horizon = 1)
}
\arguments{
\item{var}{vars::VAR or svars object}

\item{horizon}{number of steps out to calculate fe}
}
\value{
forecast error
}
\description{
Calculate the historical forecast error for a VAR out to a specified horizon
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
fe(v, 10)

}
