% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fev.R
\name{fev}
\alias{fev}
\title{Calculate the forecast error variance for a VAR out to a specified horizon}
\usage{
fev(var, h)
}
\arguments{
\item{var}{vars::VAR or svars object}

\item{h}{number of steps out to calculate fev}
}
\value{
forecast error variance
}
\description{
Calculate the forecast error variance for a VAR out to a specified horizon
}
