% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdfd.R
\name{fevfd}
\alias{fevfd}
\title{Wrapper for fevdfd that defaults to fev = TRUE.}
\usage{
fevfd(var, freqs = c(0, 2 * pi), grid_size = 1000, fev = TRUE, ...)
}
\arguments{
\item{var}{A VAR. Currently supports either a 'svars' or 'fevdvar' object.}

\item{freqs}{vector of length 2 of min and max frequencies (0:2pi)}

\item{grid_size}{how fine the grid to approximate the frequency domain}

\item{fev}{Boolean true to return fev not fevd}

\item{...}{Not currently used.}
}
\description{
Wrapper for fevdfd that defaults to fev = TRUE.
}
