% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.R
\name{hd}
\alias{hd}
\alias{hd.svars}
\title{Calculate the historical decomposition for a VAR.}
\usage{
hd(x, ...)

\method{hd}{svars}(x, ...)
}
\arguments{
\item{x}{A VAR. Currently supports either a 'svars' or 'fevdvar' object.}

\item{...}{Not currently used.}
}
\value{
Class 'fevdhd' list of historical decomposition in a data.frame.
}
\description{
Calculate the historical decomposition for a VAR.

Method to calculate hd for svars (id.chol)
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
sv <- svars::id.chol(v)
hd <- hd(sv)

}
