% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.R
\name{hist_shocks}
\alias{hist_shocks}
\title{Find the historical shocks for a VAR}
\usage{
hist_shocks(residuals, sigma)
}
\arguments{
\item{residuals}{Matrix of empirical VAR residuals}

\item{sigma}{Matrix mapping empirical residuals to structural}
}
\value{
historical shocks, matrix
}
\description{
Find the historical shocks for a VAR
}
