% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.R, R/hs.R
\name{hs}
\alias{hs}
\alias{hd.fevdvar}
\alias{hs.fevdvar}
\alias{hs.svars}
\title{Method to calculate fevdfd for fevdvar (id_fevdfd or id_fevdtd)}
\usage{
\method{hd}{fevdvar}(x, cummulative = FALSE, ...)

hs(x, ...)

\method{hs}{svars}(x, cummulative = FALSE, ...)

\method{hs}{fevdvar}(x, cummulative = FALSE, ...)
}
\arguments{
\item{x}{A VAR. Currently supports either a 'svars' or 'fevdvar' object.}

\item{cummulative}{Boolean.
Default to False, set to True for cummulative shocks.}

\item{...}{Not currently used.}
}
\value{
Class 'fevdhs' list of historical shocks in a data.frame.
}
\description{
Method to calculate fevdfd for fevdvar (id_fevdfd or id_fevdtd)

Calculate the historical structural shocks for a VAR.

Method to calculate hs for svars (id.chol)

Method to calculate hs for fevdvar (id_fevdfd or id_fevdtd)
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
sv <- svars::id.chol(v)
hs <- hs(sv)

}
