% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_fevdfd_approx.R
\name{id_fevdfd_approx}
\alias{id_fevdfd_approx}
\title{Identify...}
\usage{
id_fevdfd_approx(var, target, freqs, hmax = 1000)
}
\arguments{
\item{var, }{vars::VAR object}

\item{target, }{variable name or index to maximize its fevd}

\item{freqs}{vector of length 2 of min and max frequencies (0:2pi)}

\item{hmax}{length of irfs to calculate, longer for better approximation}
}
\value{
structural var
}
\description{
Identify...
}
