% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{id_fevdfd_bca}
\alias{id_fevdfd_bca}
\title{Translation from the Business Cycle Anatomy replication code,
used to verify the output of id_fevdfd.}
\usage{
id_fevdfd_bca(var, target, freqs)
}
\arguments{
\item{var, }{vars::VAR object}

\item{target, }{variable name or index to maximize its fevd}

\item{freqs}{vector of length 2 of min and max frequencies (0:2pi)}
}
\value{
structural var
}
\description{
Translation from the Business Cycle Anatomy replication code,
used to verify the output of id_fevdfd.
}
