% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_fevdfd.R
\name{id_fevdfd_findq}
\alias{id_fevdfd_findq}
\title{Find the rotation matrix Q that maximizes fevd over
the given frequencies for the target variable.}
\usage{
id_fevdfd_findq(betas, sigma, target_index, freqs, grid_size)
}
\arguments{
\item{betas}{matrix of coeffecients from VAR}

\item{sigma}{variance-covariance matrix from VAR}

\item{target_index}{index of variable to target}

\item{freqs}{vector of length 2 of min and max frequencies (0:pi)}

\item{grid_size}{how fine the grid to approximate the frequency domain}
}
\value{
matrix q
}
\description{
Find the rotation matrix Q that maximizes fevd over
the given frequencies for the target variable.
}
