% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_fevdtd.R
\name{id_fevdtd_findq}
\alias{id_fevdtd_findq}
\title{Find the rotation matrix Q that maximizes fevd over
the given frequencies for the target variable.}
\usage{
id_fevdtd_findq(betas, sigma, target_index, horizon)
}
\arguments{
\item{betas}{matrix of coeffecients from VAR}

\item{sigma}{variance-covariance matrix from VAR}

\item{target_index}{index of variable to target}

\item{horizon}{integer vector (can be length 1) of the horizon to maximize}
}
\value{
matrix q
}
\description{
Find the rotation matrix Q that maximizes fevd over
the given frequencies for the target variable.
}
