% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_ordered_chol.R
\name{id_ordered_chol}
\alias{id_ordered_chol}
\title{Identify a VAR with a Cholesky ordering assumption.
Assumes the given order of the variables in the VAR,
reorder the data first to get a different ordering.}
\usage{
id_ordered_chol(x)
}
\arguments{
\item{x}{a vars::VAR}
}
\value{
SVAR with cholesky id
}
\description{
Identify a VAR with a Cholesky ordering assumption.
Assumes the given order of the variables in the VAR,
reorder the data first to get a different ordering.
}
