% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\alias{irf.fevdvar}
\alias{irf.bvartools}
\title{Impulse Responses.}
\usage{
\method{irf}{fevdvar}(
  x,
  impulse = NULL,
  response = NULL,
  n.ahead = 10,
  ortho = TRUE,
  cumulative = FALSE,
  boot = TRUE,
  ci = 0.95,
  runs = 100,
  seed = NULL,
  as_matrix = FALSE,
  ...
)

\method{irf}{bvartools}(
  x,
  impulse = NULL,
  response = NULL,
  n.ahead = 10,
  ortho = TRUE,
  cumulative = FALSE,
  boot = TRUE,
  ci = 0.95,
  runs = 100,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{SVAR object of class "fevdvar".}

\item{impulse}{A character vector of the impulses, default is all variables.}

\item{response}{A character vector of the responses,
default is all variables.}

\item{n.ahead}{Integer specifying the steps.}

\item{ortho}{Not used. Here to match generic vars::irf.}

\item{cumulative}{Not used. Here to match generic vars::irf.}

\item{boot}{Not used. Here to match generic vars::irf.}

\item{ci}{Not used. Here to match generic vars::irf.}

\item{runs}{Not used. Here to match generic vars::irf.}

\item{seed}{Not used. Here to match generic vars::irf.}

\item{as_matrix}{Default to False.
Set to true to return a 3D matrix instead of a data.frame.}

\item{...}{Currently not used.}
}
\value{
Matrix of forecast error variance decomposition in frequency domain
Indexed: frequencies, variables, shocks
}
\description{
Alias for the function from svars::irf.svars,
so that the shock names are correct.
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
mvar <- id_fevdtd(v, "pi", 4:10)
vars::irf(mvar)

}
