% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf_ssv}
\alias{irf_ssv}
\title{Simple IRF function for statespace VARs}
\usage{
irf_ssv(ssv, n_ahead = 10)
}
\arguments{
\item{ssv}{Statespace VAR class.}

\item{n_ahead}{integer. How far out to calculate the impulse response.}
}
\value{
array of irfs, dimensions = c(response, shock, horizon)
}
\description{
Simple IRF function for statespace VARs
}
