% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irffd.R
\name{irffd}
\alias{irffd}
\alias{irffd.fevdvar}
\alias{irffd.svars}
\title{Calculate either the IRF in the frequency domain.}
\usage{
irffd(var, ...)

\method{irffd}{fevdvar}(var, freqs = c(0, 2 * pi), grid_size = 1000, ...)

\method{irffd}{svars}(var, freqs = c(0, 2 * pi), grid_size = 1000, ...)
}
\arguments{
\item{var}{A VAR. Currently supports either a 'svars' or 'fevdvar' object.}

\item{...}{Not currently used.}

\item{freqs}{vector of length 2 of min and max frequencies (0:2pi)}

\item{grid_size}{how fine the grid to approximate the frequency domain}
}
\value{
irffd object, storing a data.frame
}
\description{
Calculate either the IRF in the frequency domain.

Method to calculate irffd for fevdvar (this package)

Method to calculate irffd for svars (id.chol)
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
sv <- svars::id.chol(v)
irffd <- irffd(sv)

}
