% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irffd.R
\name{irffd_generic}
\alias{irffd_generic}
\title{Find the IRF in FD for a VAR}
\usage{
irffd_generic(var, freqs = c(0, 2 * pi), grid_size = 1000)
}
\arguments{
\item{var}{A VAR. Currently supports either a 'svars' or 'fevdvar' object.}

\item{freqs}{vector of length 2 of min and max frequencies (0:2pi)}

\item{grid_size}{how fine the grid to approximate the frequency domain}
}
\value{
list with vector of frequencies and matrix of irf values
}
\description{
Find the IRF in FD for a VAR
}
