% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.fevdfd}
<<<<<<< HEAD
\alias{plot.fevdvarfevd}
=======
\alias{plot.svarfevd}
>>>>>>> 483e469c89172a85f5e014d1366c1b3e1b4311f7
\title{Plot the forecast variance decomposition.}
\usage{
\method{plot}{fevdfd}(x, y, stacked = FALSE, vlines = NULL, ...)

\method{plot}{fevdvarfevd}(x, y, stacked = FALSE, vlines = NULL, ...)
}
\arguments{
\item{x}{object of class "svarfevd"}

\item{y}{Not used.}

\item{stacked}{Boolean.
True for stacked columns, False (default) for unstacked line chart.}

\item{vlines}{Vector of x-axis points at which to draw a vline.
Useful for highlighting areas.}

\item{...}{Currently not used.}
}
\value{
ggplot of fevdfd

ggplot of fevd
}
\description{
Alias for the function plot in frequency domain.
}
