% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svar_to_svar1.R
\name{svar_to_svar1}
\alias{svar_to_svar1}
\title{Convert a VAR with p lags to a representation with just one lag}
\usage{
svar_to_svar1(svar)
}
\arguments{
\item{svar}{svar object, from id.chol}
}
\value{
matrix coefficients for VAR 1 representation
}
\description{
Convert a VAR with p lags to a representation with just one lag
}
