% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fev.R
\name{fev}
\alias{fev}
\title{Calculate the forecast error variance for a VAR out to a specified horizon}
\usage{
fev(var, n_ahead = 20)
}
\arguments{
\item{var}{vars::VAR or svars object}

\item{n_ahead}{number of steps out to calculate fev}
}
\value{
forecast error variance
}
\description{
Calculate the forecast error variance for a VAR out to a specified horizon
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
mvar <- id_fevdtd(v, "pi", 4:10)
fev(mvar)

}
