% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdfd.R
\name{fevdfd}
\alias{fevdfd}
\title{Identify...}
\usage{
fevdfd(var, freqs = c(0, 2 * pi), grid_size = 1000, fev = FALSE)
}
\arguments{
\item{var, }{vars::VAR object}

\item{freqs}{vector of length 2 of min and max frequencies (0:2pi)}

\item{grid_size}{how fine the grid to approximate the frequency domain}

\item{fev}{Boolean, true to return fev not fevd}
}
\value{
forecast error variance decomposition in frequency domain
}
\description{
Identify...
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
fevd <- fevdfd(v)

}
