% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevdfd.R
\name{fevdfd}
\alias{fevdfd}
\alias{fevdfd.svars}
\alias{fevdfd.fevdvar}
\title{Calculate either the forecast error variance or fevd (default)
in the frequency domain.}
\usage{
fevdfd(var, ...)

\method{fevdfd}{svars}(var, freqs = c(0, 2 * pi), grid_size = 1000, fev = FALSE, ...)

\method{fevdfd}{fevdvar}(var, freqs = c(0, 2 * pi), grid_size = 1000, fev = FALSE, ...)
}
\arguments{
\item{var}{A VAR. Currently supports either a 'svars' or 'fevdvar' object.}

\item{...}{Not currently used.}

\item{freqs}{vector of length 2 of min and max frequencies (0:2pi)}

\item{grid_size}{how fine the grid to approximate the frequency domain}

\item{fev}{Boolean true to return fev not fevd}
}
\value{
List of forecast error variance decomposition in frequency domain
List of variables, each with a dataframe of shocks,
column 'f' are the frequencies.
}
\description{
Calculate either the forecast error variance or fevd (default)
in the frequency domain.

Method to calculate fevdfd for svars (id.chol)

Method to calculate fevdfd for svars (id.chol)
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
sv <- svars::id.chol(v)
fevd <- fevdfd(sv)

}
