% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_fevdfd.R
\name{id_fevdfd}
\alias{id_fevdfd}
\alias{id_fevdfd.varest}
\alias{id_fevdfd.varboot}
\alias{id_fevdfd.bvar}
\title{Identify the main shock by targetting the forecast error
variance contribution in the frequency domain.}
\usage{
id_fevdfd(
  x,
  target,
  freqs,
  grid_size = 1000,
  sign = "positive",
  sign_horizon = 1
)

\method{id_fevdfd}{varest}(
  x,
  target,
  freqs,
  grid_size = 1000,
  sign = "positive",
  sign_horizon = 1
)

\method{id_fevdfd}{varboot}(
  x,
  target,
  freqs,
  grid_size = 1000,
  sign = "positive",
  sign_horizon = 1
)

\method{id_fevdfd}{bvar}(
  x,
  target,
  freqs,
  grid_size = 1000,
  sign = "positive",
  sign_horizon = 1
)
}
\arguments{
\item{x}{either a \code{vars::VARS} or a \code{BVAR::bvar} object}

\item{target, }{variable name or index to maximize its fev}

\item{freqs}{vector of length 2 of min and max frequencies (0:2pi)}

\item{grid_size}{how fine the grid to approximate the frequency domain}

\item{sign}{Default to "positive". Can be "negative".  Ensures the
cummulative impact of the main shock on the target variable is the
given sign.}

\item{sign_horizon}{Default to 1. The horizon through which to accumulate the
impact of the shock.}
}
\value{
structural var
}
\description{
Identify the main shock by targetting the forecast error
variance contribution in the frequency domain.
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
mvar <- id_fevdfd(v, "pi", c(2 * pi / 32, 2 * pi / 6))

}
