% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_fevdfd.R
\name{id_fevdfd}
\alias{id_fevdfd}
\title{Identify the main shock by targetting the forecast error
variance contribution in the frequency domain.}
\usage{
id_fevdfd(var, target, freqs, grid_size = 1000)
}
\arguments{
\item{var, }{vars::VAR object}

\item{target, }{variable name or index to maximize its fev}

\item{freqs}{vector of length 2 of min and max frequencies (0:2pi)}

\item{grid_size}{how fine the grid to approximate the frequency domain}
}
\value{
structural var
}
\description{
Identify the main shock by targetting the forecast error
variance contribution in the frequency domain.
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
mvar <- id_fevdfd(v, "pi", c(2 * pi / 32, 2 * pi / 6))

}
