% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_fevdtd.R
\name{id_fevdtd}
\alias{id_fevdtd}
\title{Identify the main shock by targetting the forecast error
variance contribution in the time domain.}
\usage{
id_fevdtd(var, target, horizon)
}
\arguments{
\item{var, }{vars::VAR object}

\item{target, }{variable name or index to maximize its fevd}

\item{horizon, }{integer vector (can be length 1) of the horizon to maximize}
}
\value{
structural var
}
\description{
Identify the main shock by targetting the forecast error
variance contribution in the time domain.
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
mvar <- id_fevdtd(v, "pi", 4:10)

}
