% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_fevdtd_bca.R
\name{id_fevdtd_bca}
\alias{id_fevdtd_bca}
\title{Identify...}
\usage{
id_fevdtd_bca(var, target, horizon)
}
\arguments{
\item{var, }{vars::VAR object}

\item{target, }{variable name or index to maximize its fevd}

\item{horizon, }{integer vector (can be length 1) of the horizon to maximize}
}
\value{
structural var
}
\description{
Identify...
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
mvar <- id_fevdtd(v, "pi", 4:10)

}
