% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{id_fevdtd_bca}
\alias{id_fevdtd_bca}
\title{Translation from the Business Cycle Anatomy replication code,
used to verify the output of id_fevdtd.}
\usage{
id_fevdtd_bca(var, target, horizon)
}
\arguments{
\item{var, }{vars::VAR object}

\item{target, }{variable name or index to maximize its fevd}

\item{horizon, }{integer vector (can be length 1) of the horizon to maximize}
}
\value{
structural var
}
\description{
Translation from the Business Cycle Anatomy replication code,
used to verify the output of id_fevdtd.
}
