% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\alias{irf.fevdvar}
\title{Impulse Responses.}
\usage{
\method{irf}{fevdvar}(x, n.ahead = 20, ...)
}
\arguments{
\item{x}{SVAR object of class "fevdvar"}

\item{n.ahead}{Integer specifying the steps.}

\item{...}{Currently not used.}
}
\value{
Matrix of forecast error variance decomposition in frequency domain
Indexed: frequencies, variables, shocks
}
\description{
Alias for the function from svars::irf.svars,
so that the shock names are correct.
}
\examples{
x <- svars::USA
v <- vars::VAR(x, p = 2)
mvar <- id_fevdtd(v, "pi", 4:10)
vars::irf(mvar)

}
