% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.irffd}
\alias{plot.fevdfd}
\alias{plot.fevfd}
\alias{plot.fevdfevd}
\alias{plot.fevdirf}
\alias{plot.fevdfev}
\alias{plot.fevdhs}
\alias{plot.fevdhd}
\title{Plot the IRF in frequency domain.}
\usage{
\method{plot}{irffd}(x, y, vlines = NULL, ...)

\method{plot}{fevdfd}(x, y, stacked = TRUE, vlines = NULL, ...)

\method{plot}{fevfd}(x, y, stacked = TRUE, vlines = NULL, ...)

\method{plot}{fevdfevd}(x, y, stacked = TRUE, vlines = NULL, ...)

\method{plot}{fevdirf}(x, y = NULL, impulse_as = "colors", ...)

\method{plot}{fevdfev}(x, y = NULL, impulse_as = "colors", ...)

\method{plot}{fevdhs}(x, y = NULL, impulse_as = "colors", ...)

\method{plot}{fevdhd}(x, y = NULL, ...)
}
\arguments{
\item{x}{object of class "fevdhd"}

\item{y}{Not used.}

\item{vlines}{Vector of x-axis points at which to draw a vline.
Useful for highlighting areas.}

\item{...}{Currently not used.}

\item{stacked}{Boolean.
True for stacked columns, False (default) for unstacked line chart.}

\item{impulse_as}{Default to "colors", with all shocks on the same facet.
Other option is "cols" which puts each shock as its own column facet.}
}
\value{
ggplot of irffd

ggplot of fevdfd

ggplot of fevfd

ggplot of fevd

ggplot of irf

ggplot of fevdfev

ggplot of hs

ggplot of hd
}
\description{
Alias for the function plot in frequency domain.

Alias for the function plot in frequency domain.

Alias for the function plot in frequency domain.
}
