% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.fevdfd}
\alias{plot.svarfevd}
\title{Plot the forecast variance decomposition in frequency domain.}
\usage{
\method{plot}{fevdfd}(x, y, stacked = FALSE, vlines = NULL, ...)

\method{plot}{svarfevd}(x, y, stacked = FALSE, vlines = NULL, ...)
}
\arguments{
\item{x}{object of class "svarfevd"}

\item{y}{Not used.}

\item{stacked}{Boolean.
True for stacked columns, False (default) for unstacked line chart.}

\item{vlines}{Vector of x-axis points at which to draw a vline.
Useful for highlighting areas.}

\item{...}{Currently not used.}
}
\value{
ggplot of fevdfd

ggplot of fevdfd
}
\description{
Alias for the function plot.

Alias for the function plot.
}
